image_processor/requirements.txt
aio-pika==9.4.3
aiormq==6.8.1
alembic==1.13.3
annotated-types==0.7.0
asyncpg==0.30.0
beautifulsoup4==4.12.3
black==24.10.0
cairocffi==1.7.1
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
cssselect2==0.7.0
defusedxml==0.7.1
docopt==0.6.2
greenlet==3.1.1
grpcio==1.67.1
grpcio-tools==1.67.1
idna==3.10
Js2Py==0.74
lxml==5.3.0
Mako==1.3.6
MarkupSafe==3.0.2
multidict==6.1.0
mypy-extensions==1.0.0
packaging==24.1
pamqp==3.3.0
pathspec==0.12.1
pillow==11.0.0
platformdirs==4.3.6
propcache==0.2.0
protobuf==5.28.3
psycopg2-binary==2.9.10
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pyjsparser==2.7.1
PyPrind==2.11.3
pySmartDL==1.3.4
python-dotenv==1.0.1
pytz==2024.2
reportlab==4.2.5
requests==2.32.3
six==1.16.0
soupsieve==2.6
SQLAlchemy==2.0.36
tinycss2==1.4.0
typing_extensions==4.12.2
tzlocal==5.2
urllib3==2.2.3
webencodings==0.5.1
yarl==1.17.1
